enhancement: Start, Router and Middleware structure
