+++ backend/src/middleware/__init__.py
@@ -1,0 +1,3 @@
+from backend.src.middleware.logger_middleware import logger_middleware
+
+logger_middleware = logger_middleware
